Change types to more suitable ones, rename couple things

+++ audio_api.h
@@ -9,7 +9,7 @@ typedef struct {
     void (*write)(__int16_t *audio_buffer, long frames_in_buffer);
     void (*finish)(void);
     int buffer_size_frames;
-    long sample_rate;
+    int sample_rate;
 } audio_api_t;
 
 #define INCLUDE_AUDIO_API_H
